requirements.txt
discord
markdown
beautifulsoup4
matplotlib
numpy
bcolors
termcolors
